etsource: removed import/export distinction
let's keep things simple

+++ app/controllers/etsource/commits_controller.rb
@@ -1,6 +1,6 @@
 class Etsource::CommitsController < ApplicationController
   layout 'etsource'
-  before_filter :find_commit, :only => [:import, :export]
+  before_filter :find_commit, :only => :import
   before_filter :setup_etsource
 
   authorize_resource :class => false
@@ -21,29 +21,18 @@ class Etsource::CommitsController < ApplicationController
     @commits = @etsource.commits
   end
 
-  # like export, but will update inputs and gqueries
-  #
+  # Will import a revision to APP_CONFIG[:etsource_working_copy]
+  # and store to db gqueries and inputs
   def import
     sha = params[:id]
     @etsource.export sha
-    @commit.import! and @etsource.update_latest_import_sha(sha) and @etsource.update_latest_export_sha(sha)
+    @commit.import! and @etsource.update_latest_import_sha(sha)
     log("Import #{sha}")
     flash.now[:notice] = "Flushing ETM client cache"
     Rails.cache.clear
     # clients might need to flush their cache
     update_remote_client APP_CONFIG[:client_refresh_url]
-    restart_unicorn
-  end
-
-  # This will export a revision into APP_CONFIG[:etsource_working_copy]
-  # No changes to the db
-  #
-  def export
-    sha_id = params[:id]
-    @etsource.export(sha_id)
-    log("Use #{sha_id}")
-    restart_unicorn
-    redirect_to etsource_commits_path, :notice => "Checked out rev: #{sha_id}"
+    restart_web_server
   end
 
   private
@@ -52,8 +41,12 @@ class Etsource::CommitsController < ApplicationController
     @commit = Etsource::Commit.new(params[:id])
   end
 
-  def restart_unicorn
-    system("kill -s USR2 `cat #{Rails.root}/tmp/pids/unicorn.pid`") rescue nil
+  def restart_web_server
+    if Rails.env.production?
+      system("kill -s USR2 `cat #{Rails.root}/tmp/pids/unicorn.pid`") rescue nil
+    else
+      system "touch tmp/restart.txt"
+    end
   end
 
   def setup_etsource
